backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
apscheduler==3.10.4
pystac-client==0.8.5
sgp4==2.23
websockets==12.0
python-dotenv==1.0.1
feedparser==6.0.11
